add log interceptor
